% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{bcp}
\alias{bcp}
\title{bcp}
\usage{
bcp(df, fmla, folds = NULL, nfolds = 10, panel = c("rpart", "rf", "nnet",
  "lda", "glm", "nb", "svm"), fs = NULL, opt = NULL, ...,
  save.models = TRUE, save.formulas = TRUE, ncores = 1)
}
\arguments{
\item{nfolds}{if 'k.cv' is not null, then 'nfolds' is not used}
}
\description{
https://www.r-bloggers.com/a-brief-tour-of-the-trees-and-forests/ for description of decision trees
}
