% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{cv}
\alias{cv}
\title{Cross validation function}
\usage{
cv(x, y, nfolds = 10, folds = NULL, model_callback, predict_callback,
  fs_callback = NULL, ncores = 1)
}
\arguments{
\item{x}{a data.frame or matrix of predictors}

\item{y}{response vector}

\item{nfolds}{number of folds in cv, set k=nrow(df) to have loo}

\item{folds}{list created with caret::createFolds function}

\item{fs_callback}{number of new (training) dataset sampled with replacement from 'training'}

\item{ncores}{if ncores > 1 then it works in parallel}
}
\value{
list a (nested) list of ROCR prediction objects
}
\description{
Cross validation function
}
