% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.performance}
\alias{plot.performance}
\title{Plotting performance}
\usage{
plot.performance(classifiers, measure, x.measure = "cutoff", main = "",
  type = "l", labels = NULL, ...)
}
\arguments{
\item{classifiers}{a list, where each element contains three information
i) 'label': string representing the classifier used
ii) 'predictions': vector of numerical prediction
iii) 'truth': vector of true results}

\item{x.measure}{default 'cutoff'}
}
\description{
Plotting performance
}
