% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{plot.roc}
\alias{plot.roc}
\title{it plots ROC curves}
\usage{
plot.roc(classifiers, ...)
}
\arguments{
\item{classifiers}{a list, where each element contains three information
i) 'label': string representing the classifier used
ii) 'predictions': vector of numerical prediction
iii) 'truth': vector of true results}

\item{x.measure}{'cutoff'}
}
\description{
it plots ROC curves
}
