% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{table.performance}
\alias{table.performance}
\title{Table performance}
\usage{
table.performance(classifiers, measure, return_x_measure = FALSE,
  set_x = NULL, set_x_inequality = ">=", labels = NULL)
}
\description{
Table performance
}
