% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcp.R
\name{bcp}
\alias{bcp}
\title{bcp}
\usage{
bcp(x, y, folds = NULL, nfolds = 10, panel, save.models = TRUE,
  save.formulas = TRUE, ncores = 1)
}
\arguments{
\item{nfolds}{if 'k.cv' is not null, then 'nfolds' is not used}
}
\description{
https://www.r-bloggers.com/a-brief-tour-of-the-trees-and-forests/ for description of decision trees
}
