% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{plot.roc}
\alias{plot.roc}
\title{it plot ROC curves}
\usage{
\method{plot}{roc}(classifiers, main = "", type = "l", labels = NULL,
  ...)
}
\arguments{
\item{classifiers}{a list, where each element contains three information
i) 'label': string representing the classifier used
ii) 'predictions': vector of numerical prediction
iii) 'truth': vector of true results}
}
\description{
it plot ROC curves
}
